TCMqs_copy/requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-redis==5.2.0
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==1.1.0
kombu==5.2.4
packaging==21.3
Pillow
prompt-toolkit==3.0.29
PyMySQL
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
